// types/PostMetadata.ts
export type PostMetadata = {
  slug: string;
  title: string;
  imageUrl: string;
  thumbnailUrl: string;
  articleDate: string;
  articleReadingTime: number;
  imageAuthorName: string;
  imageAuthorUrl: string;
  imageSourceName: string;
  imageSourceUrl: string;
  excerpt: string;
  tags: string[];
  author: string;
};